+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Prevent intersecting nodes for broken links
 - Don't include [] in a link label if on new line
 - Remove lazy parsing experimental API
+- Make the GFM tagfilter extension opt-in (disabled by default) via `GFMFlavourDescriptor(useTagFilter = true)`
 
 ## [0.7.10]
 - [#211] Add support for the GFM tagfilter extension